requirements.txt
# Core
httpx>=0.26.0
pydantic>=2.5.0
pyyaml>=6.0.1
aiosqlite>=0.19.0

# AI
anthropic>=0.40.0

# API Server
fastapi>=0.109.0
uvicorn>=0.27.0

# CLI/Output
rich>=13.7.0

# Utilities
python-dateutil>=2.8.2
typing-extensions>=4.9.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Optional: Supabase
# supabase>=2.0.0
